requirements.txt
transformers==4.49.0
matplotlib
timm
pillow
peft
accelerate>=0.26.0

einops